backend/Pulse-Check/requirements.txt
amqp==5.3.1
APScheduler==3.11.2
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
django-redis==6.0.0
djangorestframework==3.17.1
kombu==5.6.2
packaging==26.2
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
